count

+++ modules/shared/models/response.go
@@ -1,9 +1,9 @@
 package models
 
 type Paging struct {
-	Page  int `json:"page"`
-	Limit int `json:"limit"`
-	Total int `json:"total"`
+	Page  int   `json:"page"`
+	Limit int   `json:"limit"`
+	Total int64 `json:"total"`
 }
 
 type Response struct {
